pyproject.toml
[project]
name = "finance-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "discord-py>=2.7.1",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.22",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.1.3",
    "litellm==1.82.6",
    "ollama>=0.6.1",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "ruff>=0.15.7",
]

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "F",  # pyflakes errors
    "B",  # bugbear warnings
    "I",  # isort (import sorting)
    "C",  # flake8-comprehensions
    "D",  # pydocstyle
    "N",  # flake8-naming
    "UP", # pyupgrade
    "TID",# flake8-tidy-imports
    "ISC",# flake8-implicit-str-concat
]

# E501은 라인 길이 초과 규칙 (format이 처리함) / # D10X pydocstyle에서 특정 규칙 무시
ignore = ["D100", "D101", "D102", "D103", "D104", "D205", "D107", "D200", "D203", "D213", "D400", "D415", "E501"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = true

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"

[tool.ruff.lint.isort]
split-on-trailing-comma=false

[tool.ruff.lint.per-file-ignores]
# 테스트 파일에서는 assert 사용 허용
"test/**" = ["S101"]
# __init__.py에서는 unused import 허용
"*/__init__.py" = ["F401"]

[tool.pytest.ini_options]
pythonpath = ["."]
